+++ js/util.js
@@ -1,0 +1,30 @@
+// Найдем случайный индекс из массива
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+// Найдем уникальный индификатор
+const createRandomIdFromRangeGenerator = (min, max) => {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      currentValue = getRandomInteger(min, max);
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+const getRandomArrayElement = (min, max) => getRandomInteger(min, max);
+
+export {createRandomIdFromRangeGenerator};
+export {getRandomArrayElement};
+export {getRandomInteger};
+
